% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsGeneral.R
\name{convert_rdata_to_csv}
\alias{convert_rdata_to_csv}
\title{convert_rdata_to_csv}
\usage{
convert_rdata_to_csv(rdata_file, output_dir = "./")
}
\arguments{
\item{rdata_file}{File in format RData}

\item{output_dir}{Destination of files converted}
}
\description{
Descripción de la función.
}
\examples{
convert_rdata_to_csv("/home/user/path_file_rdata", "/home/user/folder_files_csv/")
}
